Fix/Redo computeMainRangeFromSubranges()
Contrary to my initial assumption there are cases where we need
additional phi-defs when constructing the main liverange from subranges.
Redo the implementation in LiveRangeCalc and utilize the SSA
construction facilities there. The downside is that this is more
expensive than before, the upside is simplified easier to undertsand
code that reuses existing infrastructure for the SSA construction.

I was not able to create a testcase for this for trunk as the cases
where this happens require somewhat "independent" subranges in different
control flow parts. With the normal instruction selection those get
lowered to different vregs in my tries. I only hit these cases for some
register coalescer changes I am prepare where
computeMainRangeFromSubranges() would be called after/during coalescing.

Reviewers: qcolombet

Subscribers: llvm-commits

Differential Revision: http://reviews.llvm.org/D7887