MachineScheduler: Integrate RegisterPressure classes into PressureTracker
Instead of forcing user to either instantiate a IntervalPressure object
or a RegionPressure object and then using a reference to it in the
PressureTracker constructor, simply integrate both into a union inside
the PressureTracker removing confusion from the interface.

It may even be a good idea to simply remove the variant working with MBB
iterators, as I couldn't find any users of that.