CodeGen, AArch64, ARM, X86: Simplify SplitCSR
What is SplitCSR:
In SplitCSR callee saved registers are copied into
vregs which are added as operand to return instructions.  This is used
when callee save parameters are also used for function parameters,
because in this case the usual save in prologue/restore in epilogue
mechanisms in FrameLowering are too simplistic. Copying to a vreg lets
the register allocator track the exact liveness and omit spill/reloads
if not necessary.

Simplify SplitCSR code:
- Move generic legality checking code from
  SelectionDAG::runOnMachineFunction() to
  TargetLoweringBase::mayUseSplitCSR().
- Avoid TargetLoweringInfo::supportSplitCSR() and initializeSplitCSR()
  callbacks simply by performing the lowering as part of
  TargetLowering::LowerFormals()
- Avoid TargetLoweringInfo::insertCopiesSplitCSR() callback by using
  MachingFunction::addLiveIn() to get copies to vregs at the entry block
  and by creating CopyFromReg/CopyToReg nodes when lowering the return
  in LowerReturn. Provide a convenience function addCalleeSaveRegOps()
  for this.

Related to rdar://25638054