AsmPrinter: Emit debug information late
Attaching debug information should ideally not change the rest of the
binary. This works best when the debug information sections come last in
the file. See also the discussion in https://reviews.llvm.org/D29315 or
r197922 or rdar://15623193

This means that we should output the debug information after calling the
targets EmitEndOfAsmFile.